BROKEN conftest.py, src/add.{pt,a65}: First go at building and testing
`t8dev aslauto` is properly assembling src/add.a65. But aslauto does not
run any tests.

The top-level `conftest.py` is required for the test framework to be
able to detect the `.pt` files below the directory it's in. (This is not
an issue for tests under t8dev; it has its own conftest.py files where
it needs them.)

The test in src/add.pt fails with "fixture 'm' not found". We'll have to
look into what setup we need to do to make the testmc fixtures and
simulators available.