FAILING src/conftest.py: Import testmc.conftest to enable testmc framework
This is what gives us, e.g., the `m`, `S` and `R` fixtures and so on.

The t8dev system should also be checking to see that these sorts of
things are set up.

The test is now running, but of course failing because there's no
machine code to run yet.